controlar los tiempos para abrir el dispensador en segundos y horas

+++ src/Sonares.cpp
@@ -1,6 +1,6 @@
 #include "Sonares.h"
 
-Sonares::Sonares(int pinTrigger, int pinEcho, int maxSonar, int limit)
+Sonares::Sonares(byte pinTrigger, byte pinEcho, int maxSonar, int limit)
 {
     this->sonar = new NewPing(pinTrigger, pinEcho, maxSonar);
     this->limit = limit;
